requirements.txt
opencv-python
ultralytics
openai
python-dotenv
face_recognition
requests
numpy
